Adjust emit with args
